requirements.txt
altair==4.1.0
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
bokeh==2.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.0
entrypoints==0.3
future==0.18.2
gitdb==4.0.5
GitPython==3.1.14
idna==2.10
importlib-metadata==3.7.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
MEAutility==1.4.8
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
numexpr==2.7.3
numpy==1.20.1
packaging==20.9
pandas==1.2.3
pandastable==0.12.2.post1
pandocfilters==1.4.3
parso==0.8.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.1
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.5
ptyprocess==0.7.0
pyarrow==3.0.0
pycparser==2.20
pydeck==0.6.1
Pygments==2.8.0
pyparsing==2.4.7
pypylon==1.7.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqtgraph==0.12.0
pyrsistent==0.17.3
PySide6==6.1.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
Send2Trash==1.5.0
shiboken6==6.1.1
six==1.15.0
sklearn==0.0
smmap==3.0.5
spikecomparison==0.3.2
spikeextractors==0.9.6
spikefeatures==0.1.2
spikeinterface==0.12.0
spikemetrics==0.2.4
spikesorters==0.4.4
spiketoolkit==0.7.5
spikewidgets==0.5.3
statsmodels==0.12.2
streamlit==0.78.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tkintertable==1.3.3
tksheet==5.0.25
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
validators==0.18.2
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==2.0.1
zipp==3.4.1
